[TAS-376] refactor: 할일 추가 화면에서 카테고리 바텀 시트 컴포넌트 분리 및 제네릭 적용, 일부 디자인 수정
